fix(croppingPlanes): add gl-matrix dependancy, rename internalUtils.js
